Pierwszy CRUD - Lista zleceniodawcy

+++ src/main/java/com/sebastian/platforma/dao/IZleceniodawcaDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sebastian.platforma.domain.Zleceniodawca;
 
-public interface IZleceniodawcaDAO extends JpaRepository<Zleceniodawca, Long>{
+public interface IZleceniodawcaDAO extends JpaRepository<Zleceniodawca, Short>{
 
 }
